New Healing Machine Advancement
Removed full_heal and max_heal (dumb names, sorry), and made a (*slightly* jankily written, but fully functional) advancement which checks for you interacting with a healer with atleast one pokeball in it (any type is detected due to the stuff that makes it mildly janky looking).
A more optimized option advancement might be doable eventually, but this is the only thing I could detect to show the healer successfully being used.
Is possible to cheese it by interacting with a healer someone else is using, but shouldn't be an issue in the long run.

Updated en_us lang with new advancement, removing the prior two (the name from full_heal is kept, with a brand new description)