Updates, people, updates
Added finalized recipes for the 8 Vitamins, plus unlock recipes for them.
Removed the X-Item recipes for now, they're not needed, was only in the initial branch for critiquing and repos keep copies of that so we good.

Reorganised the Recipe Advancements folders to match the new creative tab layout, that's the majority of the changes, but it's really not anything noteworthy, just keeping it clean for future additions.

Need to look into the advancement tab background stuff, something funky's going on, but I couldn't figure it out tonight