Potentially Functional New Loot table?
I'll be honest, I don't really understand, but I fell down this hill- wait this is a commit and I'm braindying

This loot table should, in theory, work how it's outlined in the document. Hopefully.
I'll probably come back to it in future to investigate fortune mechanics, but we'll come back to that if this file doesn't break everything first